Resolved clippy output and renamed shell module folders
