+++ src/shell/ratatui/components/component.rs
