+++ src/shell/ratatui/components/input_box.rs
