+++ src/shell/ratatui/components/pokemon_table.rs
