+++ src/shell/ratatui/pages/mod.rs
