+++ src/shell/ratatui/palette.rs
